Moved LibraryScreen into a folder.
