+++ src/routes/library.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import { Image, Platform } from 'react-native';
+import { Image } from 'react-native';
 import { StackNavigator } from 'react-navigation';
 import library from 'kitsu/assets/img/tabbar_icons/library.png';
-import { LibraryScreen } from 'kitsu/screens/Library';
+import { LibraryScreen } from 'kitsu/screens/Library/LibraryScreen';
 import { navigationBarHeight, statusBarHeight } from 'kitsu/constants/app';
 import { LibraryOptions } from 'kitsu/screens/Library/LibraryOptions';
 import navigationOptions from './navigationOptions';
